requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.8.0=py36_0
alabaster=0.7.12=pypi_0
allennlp=0.7.1=pypi_0
asn1crypto=1.2.0=py36_0
astor=0.8.0=py36_0
atomicwrites=1.2.1=pypi_0
attrs=18.2.0=pypi_0
aws-xray-sdk=0.95=pypi_0
awscli=1.16.43=pypi_0
babel=2.6.0=pypi_0
blas=1.0=mkl
blessings=1.7=pypi_0
blis=0.2.4=pypi_0
bokeh=0.13.0=pypi_0
boto=2.49.0=py36_0
boto3=1.9.234=py_0
botocore=1.12.234=py_0
bz2file=0.98=py36_1
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2019.11.27=0
certifi=2019.11.28=py36_0
cffi=1.11.2=pypi_0
chardet=3.0.4=py36_1003
click=7.0=pypi_0
colorama=0.3.9=pypi_0
conllu=0.11=pypi_0
cookies=2.2.1=pypi_0
cryptography=2.8=py36h1ba5d50_0
cuda91=1.0=h4c16780_0
cudatoolkit=9.0=h13b8566_0
cudnn=7.6.4=cuda9.0_0
cupti=9.0.176=0
cycler=0.10.0=py36_0
dbus=1.13.12=h746ee38_0
distance=0.1.3=pypi_0
docker=3.5.1=pypi_0
docker-pycreds=0.3.0=pypi_0
docopt=0.6.2=pypi_0
docutils=0.15.2=py36_0
ecdsa=0.13=pypi_0
editdistance=0.5.2=pypi_0
en-core-web-sm=2.1.0=pypi_0
expat=2.2.6=he6710b0_0
faiss=1.5.3=pypi_0
fastcache=1.1.0=py36h7b6447c_0
filelock=3.0.12=pypi_0
flaky=3.4.0=pypi_0
flask=0.12.4=pypi_0
flask-cors=3.0.3=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
ftfy=5.5.0=pypi_0
future=0.17.0=pypi_0
gast=0.3.2=py_0
gensim=3.4.0=py36h14c3975_0
gevent=1.3.6=pypi_0
glib=2.63.1=h5a9c865_0
gmp=6.1.2=h6c8ec71_1
gmpy2=2.0.8=py36h10f8cd9_2
gpustat=0.6.0=pypi_0
greenlet=0.4.15=pypi_0
grpcio=1.16.1=py36hf8bcb03_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.8.0=pypi_0
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=h9c2bf20_1
idna=2.8=py36_0
imagesize=1.1.0=pypi_0
intel-openmp=2019.4=243
itsdangerous=1.1.0=pypi_0
jinja2=2.10=pypi_0
jmespath=0.9.4=py_0
joblib=0.14.0=pypi_0
jpeg=9b=h024ee3a_2
jsondiff=1.1.1=pypi_0
jsonnet=0.10.0=pypi_0
jsonpickle=1.0=pypi_0
keras-applications=1.0.7=pypi_0
keras-preprocessing=1.0.9=pypi_0
kiwisolver=1.1.0=py36he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.9.2=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
llvmlite=0.24.0=pypi_0
markdown=3.1.1=py36_0
markupsafe=1.0=pypi_0
matplotlib=2.2.3=pypi_0
mkl=2018.0.3=1
mkl_fft=1.0.6=py36h7dd41cf_0
mkl_random=1.0.1=py36h4414c95_1
mock=2.0.0=pypi_0
more-itertools=4.3.0=pypi_0
morfessor=2.0.4=py36_0
mosestokenizer=1.0.0=pypi_0
moto=1.3.4=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.1=hdf1c602_3
mpmath=1.1.0=py36_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
nltk=3.4.1=py36_0
numba=0.39.0=pypi_0
numpy=1.16.4=pypi_0
numpy-base=1.14.6=py36h81de0dd_4
numpydoc=0.8.0=pypi_0
nvidia-ml-py3=7.352.0=pypi_0
olefile=0.46=py36_0
openfile=0.0.7=pypi_0
openssl=1.1.1d=h7b6447c_3
overrides=1.9=pypi_0
packaging=17.1=pypi_0
pandas=0.25.2=py36he6710b0_0
parsimonious=0.8.0=pypi_0
pathlib=1.0.1=pypi_0
patsy=0.5.1=py36_0
pbr=5.1.0=pypi_0
pcre=8.43=he6710b0_0
pillow=6.2.1=py36h34e0f95_0
pip=19.1.1=pypi_0
pluggy=0.8.0=pypi_0
protobuf=3.9.2=py36he6710b0_0
psutil=5.6.3=pypi_0
py=1.7.0=pypi_0
pyaml=17.12.1=pypi_0
pyasn1=0.4.4=pypi_0
pycparser=2.19=py36_0
pycryptodome=3.7.0=pypi_0
pygments=2.2.0=pypi_0
pyopenssl=19.0.0=py36_0
pyparsing=2.4.2=py_0
pyqt=5.9.2=py36h05f1152_2
pysocks=1.7.1=py36_0
pytest=3.9.3=pypi_0
python=3.6.9=h265db76_0
python-dateutil=2.8.0=py36_0
python-jose=2.0.2=pypi_0
pytorch=1.1.0=py3.6_cuda9.0.176_cudnn7.5.1_0
pytz=2017.3=pypi_0
pyyaml=3.13=pypi_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
regex=2017.4.5=pypi_0
requests=2.22.0=py36_0
responses=0.10.2=pypi_0
rsa=3.4.2=pypi_0
s3transfer=0.2.1=py36_0
sacremoses=0.0.35=pypi_0
scikit-learn=0.21.3=pypi_0
scipy=1.1.0=py36hd20e5f9_0
seaborn=0.9.0=py36_0
sentencepiece=0.1.85=pypi_0
setuptools=41.6.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_0
sklearn=0.0=pypi_0
smart_open=1.8.4=py_0
snowballstemmer=1.2.1=pypi_0
spacy=2.1.6=pypi_0
sphinx=1.8.1=pypi_0
sphinxcontrib-websupport=1.1.0=pypi_0
sqlite=3.30.1=h7b6447c_0
sqlparse=0.2.4=pypi_0
srsly=0.0.7=pypi_0
statsmodels=0.10.1=py36hdd07704_0
style=1.1.0=pypi_0
sympy=1.4=py36_0
tabulate=0.8.6=pypi_0
tbb=2019.8=hfd86e86_0
tbb4py=2019.8=py36hfd86e86_0
tensorboard=1.8.0=pypi_0
tensorboardx=1.6=pypi_0
tensorflow=1.12.0=gpu_py36he68c306_0
tensorflow-base=1.12.0=gpu_py36h8e0ae2d_0
tensorflow-gpu=1.12.0=h0d30ee6_0
termcolor=1.1.0=py36_1
thinc=7.0.8=pypi_0
tk=8.6.8=hbc83047_0
tokenizers=0.0.11=pypi_0
toolwrapper=0.3.0=pypi_0
torchvision=0.2.1=pypi_0
tornado=6.0.3=py36h7b6447c_0
tqdm=4.32.1=py_0
transformers=2.4.1=pypi_0
umap=0.1.1=pypi_0
umap-learn=0.3.2=pypi_0
unidecode=1.0.22=pypi_0
update=0.0.1=pypi_0
urllib3=1.24.2=py36_0
wasabi=0.2.2=pypi_0
wcwidth=0.1.7=pypi_0
websocket-client=0.53.0=pypi_0
werkzeug=0.16.0=py_0
wheel=0.33.6=py36_0
xmltodict=0.11.0=pypi_0
xx-ent-wiki-sm=2.0.0=pypi_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
